Rewrite Bungee expansion and give it more adequate naming & more
