--- README.md
@@ -1,3 +1,0 @@
-# Bungeecord-Expansion
-Expansion for bungeecord placeholders
-[![Build Status](http://ci.extendedclip.com/buildStatus/icon?job=Bungeecord-Expansion)](http://ci.extendedclip.com/job/Bungeecord-Expansion/)
